:bug: fix(AgeField): Fixes & improvements
- Fixed responsiveness issues with the Age field
